requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pillow==11.0.0
numpy==2.2.0
requests==2.32.3
torch==2.5.1
open-clip-torch==2.29.0
